requirements.txt
tabulate
dnspython
python-whois
requests
wafw00f
